Add `cp --preserve [all]`
One common use of cp is to restore files from backups. On posix there is
the `cp -a` option. With nushell
`cp -r --preserve [ mode ownership timestamps context links xattr ]` is
something that need consoluting `cp --help` even for the short term
memory. In addition it has the disadvantage that it different for
windows as including ownership will cause an error.

This adds `cp --perserve [all]` as a short hand for preserving all the
attributes possible on the current OS.