Fix: `into string` inconsistent between float, int
When running the `cargo test` under the German locale (LANG=de_DE) one
test in `into string` fails. While debugging it, I noticed that
`into string` is inconsistent between float and int on locales that use
a different char than "." as a decimal point.

  : LANG=de_DE nu -c "(5 | into string --decimals 3) == (5.0 | into string --decimals 3)"
  false
  : LANG=de_DE nu -c "(5 | into string --decimals 3)"
  5,000
  : LANG=de_DE nu -c "(5.0 | into string --decimals 3)"
  5.000

This removes `5 | into string --decimals 3` from being locale dependent,
just like `5.0 | into string --decimals 3` currently is. As the latter
is the more common case this is the least breaking change fix. We also
don't use a localized string for the date:

  : LANG=de_DE nu -c "date now | into string"
  Wed Oct 30 11:35:44 2024

In addition, I would argue that this is the correct default case for a
scripting language. This is also what python does:

  : LANG=de_DE python3 -c 'print(3 / 4)'
  0.75

Later we could/should add an option to into string to use the locale
e.g.

  : LANG=de_DE nu -c "5 | into string --use-locale --decimals 3"
  5,000

But that shouldn't be high priority added for all supported formats at
once.

I removed the unused group_digits as that was the only other use of the
system locale.